Promote webview to EmulatorPanel class
